requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.11.0
billiard==4.1.0
boto3==1.26.27
botocore==1.29.27
celery==5.3.4
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==4.1.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==38.0.4
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==1.0.0
Django==3.2.20
django-allauth==0.54.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.13.0
django-money==3.0.0
django-otp==1.2.2
django-rest-auth==0.9.5
django-storages==1.13.1
django-templated-mail==1.1.1
django-timezone-field==6.0.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.2
MarkupSafe==2.1.1
msgpack==1.0.5
oauthlib==3.2.2
paystack==1.5.0
Pillow==9.3.0
prompt-toolkit==3.0.39
py-moneyed==2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
redis==4.6.0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
service-identity==23.1.0
sib-api-v3-sdk==7.6.0
simplejson==3.19.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stripe==5.0.0
twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
zope.interface==6.0
